% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/context.R
\name{context}
\alias{context}
\alias{current_row_info}
\alias{current_column_info}
\alias{current_n_row}
\alias{current_n_column}
\alias{row_pos}
\alias{row_rel_pos}
\alias{column_pos}
\alias{column_rel_pos}
\title{Contexts dependent functions}
\usage{
current_row_info()

current_column_info()

current_n_row()

current_n_column()

row_pos()

row_rel_pos()

column_pos()

column_rel_pos()
}
\description{
These functions are designed to work inside certain \code{matrixset} functions, to
have access to current group/matrix/row/column. Because of that, they will
not work in a general context.

The functions within which the context functions will work are \code{\link[=matrix_loop]{matrix_loop()}},
\code{\link[=row_loop]{row_loop()}}, \code{\link[=row_loop_dfl]{row_loop_dfl()}}, \code{\link[=row_loop_dfw]{row_loop_dfw()}}, \code{\link[=column_loop]{column_loop()}},
\code{\link[=column_loop_dfl]{column_loop_dfl()}} and \code{\link[=column_loop_dfw]{column_loop_dfw()}}.

Note that "current" refers to the current matrix/group/row/column, as
applicable, and possibly combined.

The context functions are:
\itemize{
\item \code{current_n_row()} and \code{current_n_column()}. They each give the number of rows
and columns, respectively, of the current matrix.

They are the context equivalent of \code{\link[=nrow]{nrow()}} and \code{\link[=ncol]{ncol()}}.
\item \code{current_row_info()} and \code{current_column_info()}. They give access to the
current row/column annotation data frame. The are the context equivlent
of \code{\link[=row_info]{row_info()}} and \code{\link[=column_info]{column_info()}}.
\item \code{row_pos()} and \code{column_pos()}. They give the current row/column indices.
The indices are the the ones before matrix subsetting.
\item \code{row_rel_pos()} and \code{column_rel_pos()}. They give the row/column indices
relative to the current matrix. They are equivalent to
\code{seq_len(current_n_row())}/\code{seq_len(current_n_column())}.
}
}
