% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn.R
\name{.mapply_column}
\alias{.mapply_column}
\title{.apply_row.row_grouped_ms <- function(.ms, ..., .matrix = NULL)
{
NextMethod()
}}
\usage{
.mapply_column(.ms, ..., .matrix = NULL, .list_input = FALSE)
}
\description{
.apply_row.row_grouped_ms <- function(.ms, ..., .matrix = NULL)
{
NextMethod()
}
}
\details{
.apply_row.col_grouped_ms <- function(.ms, ..., .matrix = NULL)
{
ans <- column_group_meta(.ms)
vals <- eval_fun(margin="row", ms=.ms, ..., matidx=.matrix,
row_first = FALSE, .simplify = FALSE,
env=rlang::caller_env())
lapply(vals, function(v) {
ans$.rows <- NULL
ans$.vals <- v
ans
})
}

.apply_row.dual_grouped_ms <- function(.ms, ..., .matrix = NULL)
{
ans <- column_group_meta(.ms)
vals <- eval_fun(margin="row", ms=.ms, ..., matidx=.matrix,
row_first = FALSE, .simplify = FALSE,
env=rlang::caller_env())
lapply(vals, function(v) {
ans$.rows <- NULL
ans$.vals <- v
ans
})
}
}
