% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mrm_plus2015}
\alias{mrm_plus2015}
\title{Table S1 and S2 of MRMPlus Paper in \code{matrixset} Format}
\format{
A \code{matrixset} of 30 rows and 45 columns
The object contains four matrices:
\describe{
\item{light_area}{Peak area of light peptides.}
\item{heavy_area}{Peak area of heavy peptides.}
\item{light_rt}{Retention time of light peptides.}
\item{heavy_rt}{Retention time of heavy peptides.}
}

The column names, analytes, are a combination of peptide sequence and
fragment ion. Rownames are the replicate names.
}
\source{
Aiyetan P, Thomas SN, Zhang Z, Zhang H. MRMPlus: an open source quality
control and assessment tool for SRM/MRM assay development. BMC Bioinformatics.
2015 Dec 12;16:411. doi: 10.1186/s12859-015-0838-z. PMID: 26652794; PMCID: PMC4676880.
}
\usage{
mrm_plus2015
}
\description{
Table S1 and S2 of MRMPlus Paper in \code{matrixset} Format
}
\keyword{datasets}
